service/requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-gitlab
requests
pydantic
chromadb
langchain
langgraph
orjson
python-dotenv
ollama